requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
boto3==1.34.162
pydantic==2.9.2
python-multipart==0.0.9
httptools==0.6.1
pytest==8.3.3
httpx==0.27.2
moto==5.0.18
Pillow==10.4.0
